bus_booking/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
razorpay
drf-spectacular
drf-spectacular-sidecar
django-cors-headers
reportlab<4
qrcode[pil]==7.4.2
pillow==10.1.0
